+++ src/lib/accessibility/FocusZone/FocusUtilities.ts
@@ -1,9 +1,9 @@
 /* tslint:disable:no-string-literal */
 
-const IS_FOCUSABLE_ATTRIBUTE = 'data-is-focusable'
-const IS_VISIBLE_ATTRIBUTE = 'data-is-visible'
-const FOCUSZONE_ID_ATTRIBUTE = 'data-focuszone-id'
-const FOCUSZONE_SUB_ATTRIBUTE = 'data-is-sub-focuszone'
+export const IS_FOCUSABLE_ATTRIBUTE = 'data-is-focusable'
+export const IS_VISIBLE_ATTRIBUTE = 'data-is-visible'
+export const FOCUSZONE_ID_ATTRIBUTE = 'data-focuszone-id'
+export const FOCUSZONE_SUB_ATTRIBUTE = 'data-is-sub-focuszone'
 
 /**
  * Gets the first focusable element.